Little refactor + added Xor in menu

+++ Wav_project/EncryptionAlgorithms.h
@@ -19,7 +19,7 @@ namespace EncryptionAlgorithms
 
 	RsaKeys generateKeys(int, int);
 
-	std::vector<int> xor (const std::vector<int>& inputVector, const std::vector<int> keyVector);
+	std::vector<short int> xor (const std::vector<short int>& inputVector, const std::vector<short int> keyVector);
 
 	std::vector<short int> encryptRsa8 (const std::vector<short int>& inputData, const short int & e, const short int & n);
 	std::vector<short int> decryptRsa8(const std::vector<short int>& inputData, const short int & e, const short int & n);
